+++ src/app/Vistas/juego/lista-juego/lista-juego.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Juego } from 'src/app/Modelos/mock-juego';
 import { FireBaseService } from 'src/app/Servicios/fire-base.service';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -20,7 +21,46 @@ export class ListaJuegoComponent {
     
     eliminaJuegos(game:Juego)
     {
-      this.fbs.deleteFireBase(game, "Juegos");
+      const swalWithBootstrapButtons = Swal.mixin({
+        customClass: {
+          confirmButton: "btn btn-success",
+          cancelButton: "btn btn-danger"
+        },
+        buttonsStyling: false
+      });
+      swalWithBootstrapButtons.fire({
+        title: "¿Estás seguro?",
+        text: "¡No se podrán revertir los cambios!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Si, eliminar!",
+        cancelButtonText: "No, cancelar!",
+        reverseButtons: true
+      }).then((result) => {
+        if (result.isConfirmed) {
+         this.fbs.deleteFireBase(game, "Juegos")
+            .then(() => swalWithBootstrapButtons.fire({
+              title: "Eliminado!",
+              text: "El juego ha sido eliminado",
+              icon: "success"
+            }))
+            .catch(() => swalWithBootstrapButtons.fire({
+              title: "Oops...!",
+              text: "El juego no ha sido eliminado",
+              icon: "error"
+            }));
+  
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire({
+            title: "Cancelado",
+            text: "El juego no ha sido eliminado",
+            icon: "error"
+          });
+        }
+      });
     }
   
 }
