+++ src/app/Vistas/usuario/lista-usuario/lista-usuario.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Usuario } from 'src/app/Modelos/mock-usuario';
 import { FireBaseService } from 'src/app/Servicios/fire-base.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-lista-usuario',
@@ -18,6 +19,46 @@ constructor(private fbs: FireBaseService) {
             .subscribe(res => this.usuarios = res);
   }
   eliminaUsuario(usuario: Usuario){
-    this.fbs.deleteFireBase(usuario, "Usuario");
+
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger"
+      },
+      buttonsStyling: false
+    });
+    swalWithBootstrapButtons.fire({
+      title: "¿Estás seguro?",
+      text: "¡No se podrán revertir los cambios!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Si, eliminar!",
+      cancelButtonText: "No, cancelar!",
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+       this.fbs.deleteFireBase(usuario, "Usuario")
+          .then(() => swalWithBootstrapButtons.fire({
+            title: "Eliminado!",
+            text: "El cliente ha sido eliminado",
+            icon: "success"
+          }))
+          .catch(() => swalWithBootstrapButtons.fire({
+            title: "Oops...!",
+            text: "El usuario no ha sido eliminado",
+            icon: "error"
+          }));
+
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire({
+          title: "Cancelado",
+          text: "El usuario no ha sido eliminado",
+          icon: "error"
+        });
+      }
+    });
   }
 }
